backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
httpx==0.27.2
feedparser==6.0.11
openai==1.51.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.9.2
